requirements.txt
xgboost~=1.4.2
torch~=1.9.0
transformers~=4.9.1
Flask==2.0.1
Flask-Cors==3.0.10
nltk~=3.6.2
tweet-preprocessor==0.6.0
gunicorn>=20.1.0
